% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_spongeffects_utility.R
\name{build_classifier_central_genes}
\alias{build_classifier_central_genes}
\title{build classifiers for central genes}
\usage{
build_classifier_central_genes(
  expression.data.set1,
  expression.data.set2,
  enrichment.modules.data.set1,
  enrichment.modules.data.set2,
  meta.data.set1,
  meta.data.set2,
  type.data.set1,
  type.data.set2,
  Metric = "Exact_match",
  tunegrid_c = c(1:100),
  n.folds = 10,
  repetitions = 3
)
}
\arguments{
\item{enrichment.modules.data.set1}{return of enrichment_modules()}

\item{enrichment.modules.data.set2}{return of enrichment_modules()}

\item{meta.data.set1}{meta data of data set 1}

\item{meta.data.set2}{meta data of data set 2}

\item{type.data.set1}{TCGA or METABRIC}

\item{type.data.set2}{TCGA or METABRIC}

\item{Metric}{metric (Exact_match, Accuracy) (default: Exact_match)}

\item{tunegrid_c}{defines the grid for the hyperparameter optimization during
cross validation (caret package) (default: 1:100)}

\item{n.folds}{number of folds to be calculated}

\item{repetitions}{number of k-fold cv iterations (default: 3)}

\item{expression_data_set1}{expression data of dataset 1,
genenames must be in rownames}

\item{expression_data_set2}{expression data of dataset 2,
genenames must be in rownames}
}
\value{
model for central genes
}
\description{
build classifiers for central genes
}
