% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_spongeffects_utility.R
\name{build_classifier_random}
\alias{build_classifier_random}
\title{build random classifiers}
\usage{
build_classifier_random(
  sponge_modules,
  train_gene_expr,
  test_gene_expr,
  train_meta_data,
  test_meta_data,
  train_meta_data_type = "TCGA",
  test_meta_data_type = "TCGA",
  metric = "Exact_match",
  tunegrid_c = c(1:100),
  n.folds = 10,
  repetitions = 3,
  min.size = 10,
  bin.size = 100,
  max.size = 200,
  min.expression = 10,
  replace = F,
  method = "OE"
)
}
\arguments{
\item{sponge_modules}{result of define_modules()}

\item{train_gene_expr}{expression data of train dataset,
genenames must be in rownames}

\item{test_gene_expr}{expression data of test dataset,
genenames must be in rownames}

\item{train_meta_data}{meta data of train dataset}

\item{test_meta_data}{meta data of test dataset}

\item{train_meta_data_type}{TCGA or METABRIC}

\item{test_meta_data_type}{TCGA or METABRIC}

\item{metric}{metric (Exact_match, Accuracy) (default: Exact_match)}

\item{tunegrid_c}{defines the grid for the hyperparameter optimization during
cross validation (caret package) (default: 1:100)}

\item{n.folds}{number of folds to be calculated}

\item{repetitions}{number of k-fold cv iterations (default: 3)}

\item{min.size}{minimum module size (default: 10)}

\item{bin.size}{bin size (default: 100)}

\item{max.size}{maximum module size (default: 200)}

\item{replace}{Possibility of keeping or removing (default) central
genes in the modules (default: T)}

\item{method}{Enrichment to be used (Overall Enrichment: OE or Gene Set
Variation Analysis: GSVA) (default: OE)}

\item{min.expr}{minimum expression (default: 10)}
}
\value{
randomized prediction model
}
\description{
build random classifiers
}
