% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ceRNA_interactions}
\alias{ceRNA_interactions}
\title{ceRNA interactions}
\format{
A data table of ceRNA interactions
typically provided by sponge
}
\usage{
data(ceRNA_interactions)
}
\description{
ceRNA interactions
}
\keyword{datasets}
