% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_spongeffects_utility.R
\name{define_modules}
\alias{define_modules}
\title{Functions to define Sponge modules, created as all the first neighbors of
the most central genes}
\usage{
define_modules(
  network,
  central.modules = F,
  remove.central = T,
  set.parallel = T,
  useCD = T
)
}
\arguments{
\item{network}{Network as dataframe and list of central nodes. First two
columns of the dataframe should contain the information of the nodes
connected by edges.}

\item{central.modules}{consider central gene as part of the module
(default: False)}

\item{remove.central}{Possibility of keeping or removing (default) central
genes in the modules (default: T)}

\item{set.parallel}{paralleling calculation of define_modules() (default: F)}
}
\value{
List of modules. Module names are the corresponding central genes.
}
\description{
Functions to define Sponge modules, created as all the first neighbors of
the most central genes
}
