% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_spongeffects_utility.R
\name{enrichment_modules}
\alias{enrichment_modules}
\title{Calculate enrichment scores}
\usage{
enrichment_modules(
  gene_expr,
  modules,
  bin.size = 100,
  min.size = 10,
  max.size = 200,
  min.expr = 10,
  method = "OE",
  cores = 1
)
}
\arguments{
\item{modules}{Result of define_modules()}

\item{bin.size}{bin size (default: 100)}

\item{min.size}{minimum module size (default: 10)}

\item{max.size}{maximum module size (default: 200)}

\item{min.expr}{minimum expression (default: 10)}

\item{method}{Enrichment to be used (Overall Enrichment: OE or Gene Set
Variation Analysis: GSVA) (default: OE)}

\item{cores}{number of cores to be used to calculate entichment scores
with gsva or ssgsea methods. Default 1}

\item{Expr.matrix}{ceRNA expression matrix}
}
\value{
matrix containing module enrichment scores (module x samples)
}
\description{
Calculate enrichment scores
}
