% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_spongeffects_utility.R
\name{fn_cal_z_score}
\alias{fn_cal_z_score}
\title{Calculate z-scores}
\usage{
fn_cal_z_score(x)
}
\arguments{
\item{x}{input to calculate z score on}
}
\value{
z score
}
\description{
Calculate z-scores
}
