% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_spongeffects_utility.R
\name{fn_combined_centrality}
\alias{fn_combined_centrality}
\title{Function to calculate centrality scores
Calculation of combined centrality scores as proposed by Del Rio et al. (2009)}
\usage{
fn_combined_centrality(CentralityMeasures)
}
\arguments{
\item{CentralityMeasures}{dataframe with centrality score measures as columns
and samples as rows}
}
\value{
Vector containing combined centrality scores
}
\description{
Function to calculate centrality scores
Calculation of combined centrality scores as proposed by Del Rio et al. (2009)
}
