% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_spongeffects_utility.R
\name{fn_convert_gene_names}
\alias{fn_convert_gene_names}
\title{Convert hugo symbols to ensemble gene names.}
\usage{
fn_convert_gene_names(
  ceRNA_expression_data,
  bioMart_gene_ensembl,
  bioMart_gene_symbol_columns
)
}
\arguments{
\item{ceRNA_expression_data}{ceRNA expression data (same structure as input
for SPONGE)
(Expression matrix (genes x samples). hugo symbols are in first column)}

\item{bioMart_gene_ensembl}{bioMart gene ensemble name
(e.g., hsapiens_gene_ensembl).
(See https://www.bioconductor.org/packages/release/bioc/vignettes/biomaRt/inst/doc/biomaRt.html)}

\item{bioMart_gene_symbol_columns}{bioMart dataset column for gene symbols
(e.g. human: hgnc_symbol, mouse: mgi_symbol)}
}
\value{
Expression matrix (genes x samples). Row names are ensembl
gene symbols
}
\description{
Convert hugo symbols to ensemble gene names.
}
