% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_spongeffects_utility.R
\name{fn_discretize_spongeffects}
\alias{fn_discretize_spongeffects}
\title{discretize}
\usage{
fn_discretize_spongeffects(v, n.cat)
}
\arguments{
\item{v}{gene distance (defined by mother function OE module function)}

\item{n.cat}{size of the bins (defined by mother function  OE module function)}
}
\value{
discretized
}
\description{
discretize
}
