% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_elastic_net.R
\name{fn_elasticnet}
\alias{fn_elasticnet}
\title{Computes an elastic net model}
\usage{
fn_elasticnet(x, y, alpha.step = 0.1)
}
\arguments{
\item{x}{miRNA expression matrix}

\item{y}{gene expression vector}

\item{alpha.step}{Step size for alpha, the tuning parameter for elastic net.}
}
\value{
The best model, i.e. the one for which the selected alpha yielded the
smallest residual sum of squares error
}
\description{
Computes an elastic net model
}
