% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_spongeffects_utility.R
\name{fn_exact_match_summary}
\alias{fn_exact_match_summary}
\title{Calibrate classification method}
\usage{
fn_exact_match_summary(data, lev = NULL, model = NULL)
}
\arguments{
\item{data}{Dataframe with module scores/covariates (modules x samples)
AND outcome variable}

\item{lev}{(default: NULL)}

\item{model}{(default: NULL)}
}
\value{
Model and confusion matrix in a list
}
\description{
Calibrate classification method
}
