% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_gene_miRNA_F_test.R
\name{fn_gene_miRNA_F_test}
\alias{fn_gene_miRNA_F_test}
\title{Perform F test for gene-miRNA elastic net model}
\usage{
fn_gene_miRNA_F_test(gene_expr, mir_expr, model, p.adj.threshold = NULL)
}
\arguments{
\item{gene_expr}{A gene expression matrix with samples in rows and
genes in columns}

\item{mir_expr}{A miRNA expression matrix with samples in rows and
genes in columns. Sample number and order has to agree with
above gene expression matrix}

\item{model}{A nested elastic net model to be tested}
}
\value{
return data frame with miRNA, fstat and adjusted p.value (BH).
}
\description{
Perform F test for gene-miRNA elastic net model
}

