% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_elastic_net.R
\name{fn_get_model_coef}
\alias{fn_get_model_coef}
\title{Extract the model coefficients from an elastic net model}
\usage{
fn_get_model_coef(model)
}
\arguments{
\item{model}{An elastic net model}
}
\value{
A data frame with miRNAs and coefficients
}
\description{
Extract the model coefficients from an elastic net model
}
