% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_elastic_net.R
\name{fn_get_rss}
\alias{fn_get_rss}
\title{Compute the residual sum of squares error for an elastic net model}
\usage{
fn_get_rss(model, x, y)
}
\arguments{
\item{model}{The elastic net model}

\item{x}{The miRNA expression}

\item{y}{The gene expression}
}
\value{
the RSS
}
\description{
Compute the residual sum of squares error for an elastic net model
}
