% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_spongeffects_utility.R
\name{fn_get_semi_random_OE}
\alias{fn_get_semi_random_OE}
\title{Function to calculate semi random enrichment scores of modules OE
(functions taken from: Jerby-Arnon et al. 2018)}
\usage{
fn_get_semi_random_OE(r, genes.dist.q, b.sign, num.rounds = 1000)
}
\arguments{
\item{r}{expression matrix}

\item{genes.dist.q}{values of the genes after binning (result of binning)}

\item{b.sign}{does the signature contain less than 2 genes?
(controll parameter) (is set by mother function (OE module function))}

\item{num.rounds}{number of rounds (default: 1000)}
}
\value{
random signature scores
}
\description{
Function to calculate semi random enrichment scores of modules OE
(functions taken from: Jerby-Arnon et al. 2018)
}
