% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_sparc.R
\name{fn_get_shared_miRNAs}
\alias{fn_get_shared_miRNAs}
\title{Identify miRNAs for which both genes have miRNA binding sites aka miRNA
response elements in the competing endogeneous RNA hypothesis}
\usage{
fn_get_shared_miRNAs(geneA, geneB, mir_interactions)
}
\arguments{
\item{geneA}{The first gene}

\item{geneB}{The second gene}

\item{mir_interactions}{A named list of genes, where for each gene all miRNA
interacting partners are listed}
}
\value{
A vector with shared RNAs of the two genes.
}
\description{
Identify miRNAs for which both genes have miRNA binding sites aka miRNA
response elements in the competing endogeneous RNA hypothesis
}
