% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_sparc.R
\name{fn_map_mimats_to_mir}
\alias{fn_map_mimats_to_mir}
\title{Helper function to map MIMAT identifers to miRNA names}
\usage{
fn_map_mimats_to_mir(mimats)
}
\arguments{
\item{mimats}{to map}
}
\value{
a vector of miRNA names
}
\description{
Helper function to map MIMAT identifers to miRNA names
}
