% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_spongeffects_utility.R
\name{fn_OE_module}
\alias{fn_OE_module}
\title{Function to calculate enrichment scores of modules OE
(functions taken from: Jerby-Arnon et al. 2018)}
\usage{
fn_OE_module(
  NormCount,
  gene.sign,
  bin.size = 100,
  num.rounds = 1000,
  set_seed = 42
)
}
\arguments{
\item{NormCount}{normalized counts}

\item{gene.sign}{significant genes}

\item{bin.size}{bin size (default: 100)}

\item{num.rounds}{number of rounds (default: 1000)}

\item{set_seed}{seed size (default: 42)}
}
\value{
Signature scores
}
\description{
Function to calculate enrichment scores of modules OE
(functions taken from: Jerby-Arnon et al. 2018)
}
