% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_gene_miRNA_regression_models.R
\name{gene_miRNA_interaction_filter}
\alias{gene_miRNA_interaction_filter}
\title{Determine miRNA-gene interactions to be considered in SPONGE}
\usage{
gene_miRNA_interaction_filter(gene_expr, mir_expr, mir_predicted_targets,
  mir_family = NULL, elastic.net = TRUE, log.every.n = 10,
  log.level = "INFO", var.threshold = 0, F.test = FALSE,
  F.test.p.adj.threshold = 0.05, coefficient.threshold = 0)
}
\arguments{
\item{gene_expr}{A matrix of gene expression}

\item{mir_expr}{A matrix of miRNA expression}

\item{mir_predicted_targets}{A data frame with miRNA in cols and genes in rows.
A 0 indicates the miRNA is not predicted to target the gene, >0 otherwise.
If this parameter is NULL all miRNA-gene interactions are tested}

\item{log.every.n}{Log only ever n iterations to limit output}

\item{log.level}{One of 'warn', 'error', 'info'}

\item{var.threshold}{Only consider genes and miRNA with
variance > var.threshold. If this parameter is NULL no variance filtering
is performed.}

\item{F.test}{If true, an F-test is performed on each model parameter to
assess its importance for the model based on the RSS of the full model vs
the RSS of the nested model without the miRNA in question. This is time
consuming and has the potential disadvantage that correlated miRNAs are
removed even though they might play a role in ceRNA interactions. Use at your
own risk.}

\item{F.test.p.adj.threshold}{If F.test is TRUE, threshold to use for
miRNAs to be included.}

\item{coefficient.threshold}{If F.test is FALSE, the
regression model considers only negative coefficients < threshold.
This is a sensible strategy since only miRNA with a negative coefficient
have an inhibitory role on gene expression.}

\item{Some}{miRNA target databases do not use mature miRNAs but miRNA family
ids since members of the same family have an identical seed sequence. To map
miRNA expression to such a database we need a mapping table. It can be
created by downloading the miR family info from the mirbase db and converting
it with the function build_mir_family_dictionary}
}
\value{
A list of genes, where for each gene, the regulating miRNA are
included as a data frame. For F.test = TRUE this is a data frame with fstat
and p-value for each miRNA. Else it is a data frame with the model
coefficients.
}
\description{
The purpose of this method is to limit the number of miRNA-gene
interactions we need to consider in SPONGE. There are 3 filtering steps:
1. variance filter (optional). Only considre genes and miRNAs with variance >
var.threshold.
2. miRNA target database filter (optional). Use a miRNA target database
provided by the user to filter for those miRNA gene interactions for which
evidence exists. This can either be predicted target interactions or
experimentally validated ones.
3. For each remaining interaction of a gene and its regulating miRNAs
use elastic net regression to achieve
a) Feature selection: We only retain miRNAs that influence gene expression
b) Effect strength: The sign of the coefficients allows
us to filter for miRNAs that down-regulate gene expression. Moreover, we can
use the coefficients to rank the miRNAs by their relative effect strength.
We strongly recommend setting up a parallel backend compatible with the
foreach package. See example and the documentation of the
foreach and doParallel packages.
}
\examples{
#library(doParallel)
#cl <- makePSOCKcluster(2)
#registerDoParallel(cl)
genes_miRNA_candidates <- gene_miRNA_interaction_filter(
gene_expr = gene_expr,
mir_expr = mir_expr,
mir_predicted_targets = NULL)
#stopCluster(cl)

#If we also perform an F-test, only few of the above miRNAs remain
genes_miRNA_candidates <- gene_miRNA_interaction_filter(
gene_expr = gene_expr,
mir_expr = mir_expr,
F.test = TRUE,
F.test.p.adj.threshold = 0.05,
mir_predicted_targets = NULL)

}
\seealso{
build_mir_family_dictionary

sponge
}

