% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_sponge.R
\name{genes_pairwise_combinations}
\alias{genes_pairwise_combinations}
\title{Compute all pairwise interactions for a number of genes as indices}
\usage{
genes_pairwise_combinations(number.of.genes)
}
\arguments{
\item{number.of.genes}{Number of genes for which all pairwise interactions
are needed}
}
\value{
data frame with one row per unique pairwise combination. To be used
as input for the sponge method.
}
\description{
Compute all pairwise interactions for a number of genes as indices
}
