% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_spongeffects_utility.R
\name{get_lncRNA_modules}
\alias{get_lncRNA_modules}
\title{prepare ceRNA network and network centralities from SPONGE / SPONGEdb}
\usage{
get_lncRNA_modules(
  bioMart_gene_ensembl,
  weighted_node_centrality,
  cutoff = 1000
)
}
\arguments{
\item{bioMart_gene_ensembl}{bioMart gene ensemble name
(e.g., hsapiens_gene_ensembl).
(See https://www.bioconductor.org/packages/release/bioc/vignettes/biomaRt/inst/doc/biomaRt.html)}

\item{weighted_node_centrality}{output from filter_ceRNA_network()}

\item{cutoff}{the top cutoff modules will be returned (default: 1000)}
}
\value{
top cutoff modules, with lncRNA as central genes
}
\description{
prepare ceRNA network and network centralities from SPONGE / SPONGEdb
}
