% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mir_family_info}
\alias{mir_family_info}
\title{miRNA family mapping table}
\format{A data frame used to map miR.family ids to MiRBase.Accession ids}
\source{
ftp://mirbase.org/pub/mirbase/CURRENT/miFam.dat.gz
}
\usage{
mir_family_info
}
\description{
miRNA family mapping table
}
\keyword{datasets}

