% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mircode_symbol}
\alias{mircode_symbol}
\title{mircode predicted miRNA gene interactions}
\format{
A matrix gene symbols vs miRNA family names. >=1 if interaction is
predicted, 0 otherwise
}
\source{
http://www.mircode.org/download.php
}
\description{
mircode predicted miRNA gene interactions
}
\keyword{datasets}
