% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_spongeffects_utility.R
\name{plot_heatmaps}
\alias{plot_heatmaps}
\title{plots the heatmaps from training_and_test_model
(see Boniolo and Hoffmann 2022 et al. Fig. 6)}
\usage{
plot_heatmaps(
  trained_model,
  spongEffects,
  meta_data,
  label,
  sampleIDs,
  Modules_to_Plot = 10,
  show.rownames = T,
  show.colnames = F,
  title
)
}
\arguments{
\item{trained_model}{returned from train_and_test_model}

\item{spongEffects}{output of enrichment_modules()}

\item{meta_data}{metadata of samples
(retrieved from prepare_tcga_for_spongEffects() or
from prepare_metabric_for_spongEffects())}

\item{label}{Column of metadata to use as label in classification model}

\item{sampleIDs}{Column of metadata containing sample/patient IDs to be matched with column names of spongEffects scores}

\item{Modules_to_Plot}{Number of modules to plot in the heatmap. Default = 2}

\item{show.rownames}{Add row names (i.e. module names) to the heatmap. Default = F}

\item{show.colnames}{Add column names (i.e. sample names) to the heatmap. Default = F}
}
\value{
ComplexHeatmap object
NOT FUNCTIONAL
}
\description{
plots the heatmaps from training_and_test_model
(see Boniolo and Hoffmann 2022 et al. Fig. 6)
}
