% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_spongeffects_utility.R
\name{plot_heatmaps_training_test}
\alias{plot_heatmaps_training_test}
\title{plots the heatmaps from training_and_test_model
(see Boniolo and Hoffmann 2022 et al. Fig. 6)}
\usage{
plot_heatmaps_training_test(
  trained_model,
  sponge_modules,
  enrichment_modules,
  meta_data,
  meta_data_type = "TCGA",
  subtypes,
  bioMart_gene_symbol_columns = "hgnc_symbol",
  bioMart_gene_ensembl = "hsapiens_gene_ensembl",
  met_brewer.palette_subtypes = "Austria",
  met_brewer.palette_heatmap = "Troy"
)
}
\arguments{
\item{trained_model}{returned from train_and_test_model}

\item{sponge_modules}{result of define_modules()}

\item{enrichment_modules}{return of enrichment_modules()}

\item{meta_data}{meta data}

\item{meta_data_type}{TCGA or METABRIC}

\item{subtypes}{array of subtypes
(e.g., c("Normal", "LumA", "LumB", "Her2", "Basal"))}

\item{bioMart_gene_symbol_columns}{bioMart dataset column for gene symbols
(e.g. human: hgnc_symbol, mouse: mgi_symbol)
(default: hgnc_symbol)}

\item{bioMart_gene_ensembl}{bioMart gene ensemble name
(e.g., hsapiens_gene_ensembl).}

\item{met_brewer.palette_subtypes}{which met brewer colour palette should be
used for the subtypes (default: "Austria")}

\item{met_brewer.palette_heatmap}{which  met brewer colour palette should be
used for the heatmap (default "Troy")}
}
\value{
NOT FUNCTIONAL
}
\description{
plots the heatmaps from training_and_test_model
(see Boniolo and Hoffmann 2022 et al. Fig. 6)
}
