% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_spongeffects_utility.R
\name{plot_miRNAs_per_gene}
\alias{plot_miRNAs_per_gene}
\title{plot expression of miRNAs in module for specific RNA (no negative values!)}
\usage{
plot_miRNAs_per_gene(
  target,
  genes_miRNA_candidates,
  mi_rna_expr,
  meta,
  log_transform = T,
  pseudocount = 0.001,
  unit = "counts"
)
}
\arguments{
\item{target}{the target gene}

\item{genes_miRNA_candidates}{the miRNAs of the module to plot the relative expression against}

\item{mi_rna_expr}{the miRNA expression}

\item{meta}{the meta information for annotations}

\item{unit}{the unit to display for the data. This parameter does not chance the behaviour of the method and the data is expected to match the provided unit}

\item{log_trasform}{defaut T: whether the results should be log transformed}

\item{pseudocout}{defaut 1e-3: the pseudocounts}
}
\value{
plot object
}
\description{
plot expression of miRNAs in module for specific RNA (no negative values!)
}
