% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_spongeffects_utility.R
\name{plot_target_gene_expressions}
\alias{plot_target_gene_expressions}
\title{plot the target gene heatmap}
\usage{
plot_target_gene_expressions(
  target,
  target_genes,
  gene_expression,
  meta,
  gtf_raw = NA,
  annotation = NA
)
}
\arguments{
\item{target}{the target gene}

\item{target_genes}{the genes of the module to plot the heatmap with}

\item{gene_expression}{the gene expression}

\item{meta}{the meta information for annotations}

\item{gtf_raw}{defaut NA: if given, will convert to hgnc}

\item{annotation}{defaut NA: gene annotations}
}
\value{
pheatmap heatmap
}
\description{
plot the target gene heatmap
}
