% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_spongeffects_utility.R
\name{Random_spongEffects}
\alias{Random_spongEffects}
\title{Define random modules}
\usage{
Random_spongEffects(
  sponge_modules,
  gene_expr,
  min.size = 10,
  bin.size = 100,
  max.size = 200,
  min.expression = 10,
  replace = F,
  method = "OE",
  cores = 1
)
}
\arguments{
\item{sponge_modules}{result of define_modules()}

\item{gene_expr}{Input expression matri}

\item{min.size}{minimum module size (default: 10)}

\item{bin.size}{bin size (default: 100)}

\item{max.size}{maximum module size (default: 200)}

\item{min.expression}{minimum expression (default: 10)}

\item{replace}{Possibility of keeping or removing (default) central
genes in the modules (default: F)}

\item{method}{Enrichment to be used (Overall Enrichment: OE or Gene Set
Variation Analysis: GSVA) (default: OE)}

\item{cores}{number of cores to be used to calculate entichment scores with gsva or ssgsea methods. Default 1}
}
\value{
A list with randomly defined modules and related enrichment scores
}
\description{
Define random modules
}
