% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_simulation.R
\name{sample_zero_mmscor_data}
\alias{sample_zero_mmscor_data}
\title{Sample mmscor coefficients from pre-computed covariance matrices}
\usage{
sample_zero_mmscor_data(cov.matrices, number.of.samples = 100,
  number.of.datasets = 100)
}
\arguments{
\item{cov.matrices}{a list of pre-computed covariance matrices}

\item{number.of.samples}{the number of samples available in the expression
data}

\item{number.of.datasets}{the number of mmscor coefficients to be sampled
from each covariance matrix}
}
\value{
a vector of mmscor coefficients
}
\description{
Sample mmscor coefficients from pre-computed covariance matrices
}
\examples{
#we select from the pre-computed covariance matrices in SPONGE
those 10 for m = 5 miRNAs and gene-gene correlation 0.6
cov_matrices_selected <- cov.matrices[["5"]][["0.6"]]
sample_zero_mmscor_data(cov.matrices = cov_matrices_selected,
number.of.samples = 200, number.of.datasets = 10)
}
\seealso{
sample_zero_mmscor_cov
}

