% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_imports.R
\docType{package}
\name{SPONGE-package}
\alias{SPONGE}
\alias{SPONGE-package}
\title{Package shared imports}
\description{
Centralized imports for functions used across multiple files.
These roxygen directives generate `NAMESPACE` entries via roxygen2.
}
\author{
\strong{Maintainer}: Markus List \email{markus.list@tum.de} (\href{https://orcid.org/0000-0002-0941-4168}{ORCID})

Authors:
\itemize{
  \item Markus List \email{markus.list@tum.de} (\href{https://orcid.org/0000-0002-0941-4168}{ORCID})
  \item Markus Hoffmann \email{markus.hoffmann@nih.gov} (\href{https://orcid.org/0000-0002-1920-288X}{ORCID})
  \item Lena Strasser \email{lena.j.strasser@tum.de} (\href{https://orcid.org/0009-0007-7881-6818}{ORCID})
  \item Fabio Boniolo
  \item Azim Dehghani Amirabad
  \item Dennis Kostka
  \item Marcel H. Schulz
}

}
\keyword{internal}
