% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_plots.R
\name{sponge_network}
\alias{sponge_network}
\title{Prepare a sponge network for plotting}
\usage{
sponge_network(sponge_result, mir_data, target.genes = NULL,
  show.sponge.interaction = TRUE, show.mirnas = "none",
  min.interactions = 3)
}
\arguments{
\item{sponge_result}{ceRNA interactions as produced by the sponge method.}

\item{mir_data}{miRNA interactions as produced by sponge_gene_miRNA_interaction_filter}

\item{target.genes}{a character vector  to select a subset of genes}

\item{show.sponge.interaction}{whether to connect ceRNAs}

\item{show.mirnas}{one of none, shared, all}

\item{min.interactions}{minimum degree of a gene to be shown}
}
\value{
a list of nodes and edges
}
\description{
Prepare a sponge network for plotting
}
\examples{
sponge_network(ceRNA_interactions, mir_interactions)
}
