% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_plots.R
\name{sponge_plot_network}
\alias{sponge_plot_network}
\title{Plot a sponge network}
\usage{
sponge_plot_network(sponge_result, mir_data,
  layout = "layout.fruchterman.reingold", force.directed = FALSE, ...)
}
\arguments{
\item{sponge_result}{ceRNA interactions as produced by the sponge method.}

\item{mir_data}{miRNA interactions as produced by sponge_gene_miRNA_interaction_filter}

\item{layout}{one of the layout methods supported in the visNetwork package}

\item{force.directed}{whether to produce a force directed network, gets
slow for large networks}

\item{...}{further params for sponge_network}
}
\value{
shows a plot
}
\description{
Plot a sponge network
}
\examples{
#sponge_plot_network(ceRNA_interactions, mir_interactions)
}
