% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_network_centralities.R
\name{sponge_plot_network_centralities}
\alias{sponge_plot_network_centralities}
\title{plot node network centralities}
\usage{
sponge_plot_network_centralities(network_centralities, measure = "all",
  x = "degree", top = 5, base_size = 18)
}
\arguments{
\item{network_centralities}{a result from sponge_node_centralities()}

\item{measure}{one of 'all', 'degree', 'ev' or 'btw'}

\item{x}{plot against another column in the data table, defaults to degree}

\item{top}{label the top x samples in the plot}

\item{base_size}{size of the text in the plot}
}
\value{
a plot
}
\description{
plot node network centralities
}
\examples{
#sponge_plot_network_centralities
}
