% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_plot_simulation_results.R
\name{sponge_plot_simulation_results}
\alias{sponge_plot_simulation_results}
\title{Plot simulation results for different null models}
\usage{
sponge_plot_simulation_results(null_model_data)
}
\arguments{
\item{null_model_data}{the output of sponge_build_null_model}
}
\description{
Plot simulation results for different null models
}
\examples{
sponge_plot_simulation_results(sponge_build_null_model(100, 100))
}
