% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_benchmark.R
\name{sponge_run_benchmark}
\alias{sponge_run_benchmark}
\title{run sponge benchmark}
\usage{
sponge_run_benchmark(gene_expr, mir_expr, mir_interactions,
  num_of_genes_to_test = c(25), folder = "")
}
\arguments{
\item{gene_expr}{a gene expression matrix}

\item{mir_expr}{a miRNA expression matrix}

\item{mir_interactions}{a list of mir interaction sources such as targetscan, mircode, etc.}

\item{num_of_genes_to_test}{a vector of numbers of genes to be tested, e.g. c(250,500)}

\item{folder}{where the results should be saved}
}
\value{
a list (regression, no regression) of lists (single miRNA, pooled miRNAs) of benchmark results
}
\description{
run sponge benchmark
}

