% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_cancer_mir_expr}
\alias{test_cancer_mir_expr}
\title{example test miRNA data for spongEffects}
\format{
a matrix with miRNA expression data
}
\usage{
data(test_cancer_mir_expr)
}
\description{
example test miRNA data for spongEffects
}
\keyword{datasets}
