% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_spongeffects_utility.R
\name{train_and_test_model}
\alias{train_and_test_model}
\title{tests and trains a model for a disease using a training and test data set
(e.g., TCGA-BRCA and METABRIC)}
\usage{
train_and_test_model(
  Modules_training,
  Modules_training.metadata,
  Modules_testing,
  Modules_testing.metadata,
  Modules_testing.metadata.type,
  Metric = "Exact_match",
  tunegrid_c = c(1:100),
  n_folds = 10,
  repetitions = 3
)
}
\arguments{
\item{Modules_training}{return from enrichment_modules() function}

\item{Modules_training.metadata}{TCGA-BRCA data}

\item{Modules_testing}{return from enrichment_modules() function}

\item{Modules_testing.metadata}{e.g., METABRIC data or other TCGA data}

\item{Modules_testing.metadata.type}{e.g. METABRIC or TCGA}

\item{Metric}{metric (Exact_match, Accuracy) (default: Exact_match)}

\item{tunegrid_c}{defines the grid for the hyperparameter optimization during
cross validation (caret package) (default: 1:100)}

\item{n_folds}{number of folds (default: 10)}

\item{repetitions}{number of k-fold cv iterations (default: 3)}
}
\value{
returns a list with the trained model and the prediction results
}
\description{
tests and trains a model for a disease using a training and test data set
(e.g., TCGA-BRCA and METABRIC)
}
