% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{train_cancer_gene_expr}
\alias{train_cancer_gene_expr}
\title{example training expression data for spongEffects}
\format{
a matrix with gene expression data
}
\usage{
data(train_cancer_gene_expr)
}
\description{
example training expression data for spongEffects
}
\keyword{datasets}
