% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{train_cancer_mir_expr}
\alias{train_cancer_mir_expr}
\title{example training miRNA data for spongEffects}
\format{
a matrix with miRNA expression data
}
\usage{
data(train_cancer_mir_expr)
}
\description{
example training miRNA data for spongEffects
}
\keyword{datasets}
