% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{train_ceRNA_interactions}
\alias{train_ceRNA_interactions}
\title{example train ceRNA interactions for spongEffects}
\format{
(obtained by SPONGE method)
}
\usage{
data(train_ceRNA_interactions)
}
\description{
example train ceRNA interactions for spongEffects
}
\keyword{datasets}
