% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{train_genes_miRNA_candidates}
\alias{train_genes_miRNA_candidates}
\title{example candidate genes for miRNA filtering}
\format{
A data frame with candidate genes and associated miRNA scores
}
\usage{
data(train_genes_miRNA_candidates)
}
\description{
example candidate genes for miRNA filtering
}
\keyword{datasets}
