% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{train_network_centralities}
\alias{train_network_centralities}
\title{example train network centralities for spongEffects}
\format{
(obtained by SPONGE method)
}
\usage{
data(train_network_centralities)
}
\description{
example train network centralities for spongEffects
}
\keyword{datasets}
