% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_spongeffects_utility.R
\name{calibrate_model}
\alias{calibrate_model}
\title{tests and trains a model for a disease using a training and test data set
(e.g., TCGA-BRCA and METABRIC)}
\usage{
calibrate_model(
  modules,
  modules_metadata,
  label,
  sampleIDs,
  Metric = "Exact_match",
  tunegrid_c = c(1:100),
  n_folds = 10,
  repetitions = 3
)
}
\arguments{
\item{modules}{return from enrichment_modules() function}

\item{modules_metadata}{metadata table containing information about samples/patients}

\item{label}{Column of metadata to use as label in classification model}

\item{sampleIDs}{Column of metadata containing sample/patient IDs to be matched with column names of spongEffects scores}

\item{Metric}{metric (Exact_match, Accuracy) (default: Exact_match)}

\item{tunegrid_c}{defines the grid for the hyperparameter optimization during
cross validation (caret package) (default: 1:100)}

\item{n_folds}{number of folds (default: 10)}

\item{repetitions}{number of k-fold cv iterations (default: 3)}
}
\value{
returns a list with the trained model and the prediction results
Calibrate classification RF classification model

returns a list with the trained model and the prediction results
}
\description{
tests and trains a model for a disease using a training and test data set
(e.g., TCGA-BRCA and METABRIC)
}
