% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_check_and_convert_expr_data.R
\name{check_and_convert_expression_data}
\alias{check_and_convert_expression_data}
\title{Checks if expression data is in matrix or ExpressionSet format and
converts the latter to a standard matrix.}
\usage{
check_and_convert_expression_data(expr_data)
}
\arguments{
\item{expr_data}{expr_data as matrix or ExpressionSet}
}
\value{
expr_data as matrix
}
\description{
Checks if expression data is in matrix or ExpressionSet format and
converts the latter to a standard matrix.
}
\examples{
check_and_convert_expression_data(gene_expr)
}
