% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_spongeffects_utility.R
\name{enrichment_modules}
\alias{enrichment_modules}
\title{Calculate enrichment scores}
\usage{
enrichment_modules(
  Expr.matrix,
  modules,
  bin.size = 100,
  min.size = 10,
  max.size = 200,
  method = "OE"
)
}
\arguments{
\item{Expr.matrix}{ceRNA expression matrix}

\item{modules}{Result of define_modules()}

\item{bin.size}{bin size (default: 100)}

\item{min.size}{minimum module size (default: 10)}

\item{max.size}{maximum module size (default: 200)}

\item{method}{Enrichment to be used (Overall Enrichment: OE or Gene Set
Variation Analysis: GSVA) (default: OE)}
}
\value{
matrix containing module enrichment scores (module x samples)
}
\description{
Calculate enrichment scores
}
