% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ensembl.df}
\alias{ensembl.df}
\title{example potential central nodes}
\format{
(downloaded via biomaRt)
}
\usage{
ensembl.df
}
\description{
example potential central nodes
}
\keyword{datasets}
