% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_spongeffects_utility.R
\name{fn_filter_network}
\alias{fn_filter_network}
\title{#' Calculate z-scores
#'
#' @import tidyverse
#' @import caret
#' @import dplyr
#' @import Biobase
#' @import biomaRt
#' @import randomForest
#' @import ggridges
#' @import MetBrewer
#' @import cvms
#' @import ComplexHeatmap
#' @import miRBaseConverter
#'
#' @param x input to calculate z score on
#'
#' @return z score
fn_cal_z_score <- function(x){
  (x - mean(x)) / sd(x)
}
Preprocessing ceRNA network}
\usage{
fn_filter_network(network, mscor.threshold = 0.1, padj.threshold = 0.01)
}
\arguments{
\item{network}{ceRNA network as data (typically present in the outputs of
sponge)}

\item{mscor.threshold}{mscor threshold (default 0.1)}

\item{padj.threshold}{adjusted p-value threshold (default 0.01)}
}
\value{
filtered ceRNA network
}
\description{
#' Calculate z-scores
#'
#' @import tidyverse
#' @import caret
#' @import dplyr
#' @import Biobase
#' @import biomaRt
#' @import randomForest
#' @import ggridges
#' @import MetBrewer
#' @import cvms
#' @import ComplexHeatmap
#' @import miRBaseConverter
#'
#' @param x input to calculate z score on
#'
#' @return z score
fn_cal_z_score <- function(x){
  (x - mean(x)) / sd(x)
}
Preprocessing ceRNA network
}
