% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_spongeffects_utility.R
\name{fn_RF_classifier}
\alias{fn_RF_classifier}
\title{RF classification model}
\usage{
fn_RF_classifier(
  Input.object,
  K,
  rep,
  metric = "Exact_match",
  tunegrid,
  set_seed = 42
)
}
\arguments{
\item{Input.object}{data.frame made by predictors and dependent variable}

\item{K}{number of folds (k-fold)}

\item{rep}{number of times repeating the cross validation}

\item{metric}{metric (Exact_match, Accuracy) (default: Exact_match)}

\item{tunegrid}{defines the grid for the hyperparameter optimization during
cross validation (caret package)}

\item{set_seed}{set seed (default: 42)}
}
\description{
RF classification model
}
