% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_expr}
\alias{gene_expr}
\title{Gene expression test data set}
\format{
A data frame of expression values
with samples in columns and genes in rows
}
\usage{
gene_expr
}
\description{
Gene expression test data set
}
\keyword{datasets}
