% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mir_expr}
\alias{mir_expr}
\title{miRNA expression test data set}
\format{A data frame of expression values
with samples in columns and miRNA in rows}
\usage{
mir_expr
}
\description{
miRNA expression test data set
}
\keyword{datasets}

