% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mir_interactions}
\alias{mir_interactions}
\title{miRNA / gene interactions}
\format{A data frame of regression coefficients
typically provided by sponge_gene_miRNA_interaction_filter}
\usage{
mir_interactions
}
\description{
miRNA / gene interactions
}
\keyword{datasets}
