% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_spongeffects_utility.R
\name{plot_accuracy_sensitivity_specificity}
\alias{plot_accuracy_sensitivity_specificity}
\title{list of plots for (1) accuracy and (2) sensitivity + specificity
(see Boniolo 2022 et al. Fig. 3a and Fig. 3b)}
\usage{
plot_accuracy_sensitivity_specificity(
  trained.model,
  CentralGenes.model,
  Random.model,
  training_dataset_name,
  testing_dataset_name,
  subtypes
)
}
\arguments{
\item{trained.model}{returned from train_and_test_model}

\item{CentralGenes.model}{returned from build_classifier_central_genes()}

\item{training_dataset_name}{name of training (e.g., TCGA)}

\item{testing_dataset_name}{name of testing set (e.g., METABRIC)}

\item{subtypes}{array of subtypes
(e.g., c("Normal", "LumA", "LumB", "Her2", "Basal"))}

\item{random.model}{returned from train_and_test_model using the randomization}
}
\value{
list of plots for (1) accuracy and (2) sensitivity + specificity
}
\description{
list of plots for (1) accuracy and (2) sensitivity + specificity
(see Boniolo 2022 et al. Fig. 3a and Fig. 3b)
}
