% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_spongeffects_utility.R
\name{plot_density_scores}
\alias{plot_density_scores}
\title{plots the density of the model scores for subtypes (see Boniolo and Hoffmann 2022 et al. Fig. 2)}
\usage{
plot_density_scores(
  trained_model,
  modules,
  meta_data,
  meta_data_type,
  subtypes,
  bioMart_gene_symbol_columns = "hgnc_symbol",
  bioMart_gene_ensembl = "hsapiens_gene_ensembl"
)
}
\arguments{
\item{trained_model}{returned from train_and_test_model}

\item{modules}{output of enrichment_modules()}

\item{meta_data}{metadata of samples
(retrieved from prepare_tcga_for_spongEffects() or
from prepare_metabric_for_spongEffects())}

\item{meta_data_type}{TCGA or METABRIC}

\item{subtypes}{array of subtypes
(e.g., c("Normal", "LumA", "LumB", "Her2", "Basal"))}

\item{bioMart_gene_symbol_columns}{bioMart dataset column for gene symbols
(e.g. human: hgnc_symbol, mouse: mgi_symbol)
(default: hgnc_symbol)}

\item{bioMart_gene_ensembl}{bioMart gene ensemble name
(e.g., hsapiens_gene_ensembl).}
}
\value{
plots density scores for subtypes
}
\description{
plots the density of the model scores for subtypes (see Boniolo and Hoffmann 2022 et al. Fig. 2)
}
