% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_spongeffects_utility.R
\name{plot_density_scores}
\alias{plot_density_scores}
\title{plots the density of the model scores for subtypes (see Boniolo and Hoffmann 2022 et al. Fig. 2)}
\usage{
plot_density_scores(trained_model, spongEffects, meta_data, label, sampleIDs)
}
\arguments{
\item{trained_model}{returned from train_and_test_model}

\item{spongEffects}{output of enrichment_modules()}

\item{meta_data}{metadata of samples
(retrieved from prepare_tcga_for_spongEffects() or
from prepare_metabric_for_spongEffects())}

\item{label}{Column of metadata to use as label in classification model}

\item{sampleIDs}{Column of metadata containing sample/patient IDs to be matched with column names of spongEffects scores}

\item{meta_data_type}{TCGA or METABRIC}
}
\value{
plots density scores for subtypes
}
\description{
plots the density of the model scores for subtypes (see Boniolo and Hoffmann 2022 et al. Fig. 2)
}
