% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_spongeffects_utility.R
\name{plot_heatmaps}
\alias{plot_heatmaps}
\title{}
\usage{
plot_heatmaps(
  trained_model,
  spongEffects,
  meta_data,
  label,
  sampleIDs,
  Modules_to_Plot = 2,
  show.rownames = F,
  show.colnames = F
)
}
\arguments{
\item{trained_model}{returned from train_and_test_model}

\item{spongEffects}{output of enrichment_modules()}

\item{meta_data}{metadata of samples
(retrieved from prepare_tcga_for_spongEffects() or
from prepare_metabric_for_spongEffects())}

\item{label}{Column of metadata to use as label in classification model}

\item{sampleIDs}{Column of metadata containing sample/patient IDs to be matched with column names of spongEffects scores}

\item{Modules_to_Plot}{Number of modules to plot in the heatmap. Default = 2}

\item{show.rownames}{Add row names (i.e. module names) to the heatmap. Default = F}

\item{show.colnames}{Add column names (i.e. sample names) to the heatmap. Default = F}
}
\value{
ComplexHeatmap object
NOT FUNCTIONAL
}
\description{
trained.model<-trained_model
    Sponge.modules<-sponge_modules
    enrichment.modules<-enrichment_modules
    meta.data<-meta_data
    type.data.set<-meta_data_type
}
\details{

}
