% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_spongeffects_utility.R
\name{plot_involved_miRNAs_to_modules}
\alias{plot_involved_miRNAs_to_modules}
\title{plots the heatmap of miRNAs invovled in the interactions of the modules
(see Boniolo 2022 et al. Fig. 7a)}
\usage{
plot_involved_miRNAs_to_modules(
  Sponge.modules,
  trained.model,
  dir_miRNAs_significance,
  k_modules = 25,
  filter_miRNAs = 3,
  bioMart_gene_symbol_columns = "hgnc_symbol",
  bioMart_gene_ensembl = "hsapiens_gene_ensembl"
)
}
\arguments{
\item{Sponge.modules}{result of define_modules()}

\item{trained.model}{returned from train_and_test_model}

\item{dir_miRNAs_significance}{output of SPONGE or SPONGEdb (RDATA object)}

\item{k_modules}{top k modules to be shown (default: 25)}

\item{filter_miRNAs}{min rowsum to be reach of miRNAs (default: 3.0)}

\item{bioMart_gene_symbol_columns}{bioMart dataset column for gene symbols
(e.g. human: hgnc_symbol, mouse: mgi_symbol)
(default: hgnc_symbol)}

\item{bioMart_gene_ensembl}{bioMart gene ensemble name
(e.g., hsapiens_gene_ensembl).}
}
\value{
plot object
}
\description{
plots the heatmap of miRNAs invovled in the interactions of the modules
(see Boniolo 2022 et al. Fig. 7a)
}
