% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_spongeffects_utility.R
\name{plot_top_modules}
\alias{plot_top_modules}
\title{plots the top x gini index modules (see Boniolo and Hoffmann 2022 et al. Figure 5)}
\usage{
plot_top_modules(
  trained_model,
  k_modules = 25,
  k_modules_red = 10,
  text_size = 16,
  bioMart_gene_symbol_columns = "hgnc_symbol",
  bioMart_gene_ensembl = "hsapiens_gene_ensembl"
)
}
\arguments{
\item{trained_model}{returned from train_and_test_model}

\item{k_modules}{top k modules to be shown (default: 25)}

\item{k_modules_red}{top k modules shown in red - NOTE: must be smaller
than k_modules (default: 10)}

\item{text_size}{text size (default 16)}

\item{bioMart_gene_symbol_columns}{bioMart dataset column for gene symbols
(e.g. human: hgnc_symbol, mouse: mgi_symbol)
(default: hgnc_symbol)}

\item{bioMart_gene_ensembl}{bioMart gene ensemble name
(e.g., hsapiens_gene_ensembl).}
}
\value{
plot object for lollipop plot
}
\description{
plots the top x gini index modules (see Boniolo and Hoffmann 2022 et al. Figure 5)
}
