% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_significance.R
\name{sponge_build_null_model}
\alias{sponge_build_null_model}
\title{Build null model for p-value computation}
\usage{
sponge_build_null_model(number.of.datasets = 1e+05, number.of.samples)
}
\arguments{
\item{number.of.datasets}{the number of datesets defining the precision of the p-value}

\item{number.of.samples}{the number of samples in the expression data}
}
\value{
a list (for various values of m) of lists (for various values of k)
of lists of simulated data sets, drawn from a set of precomputed covariance matrices
}
\description{
Build null model for p-value computation
}
\examples{
sponge_build_null_model(100, 100)
}

