% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_network_centralities.R
\name{sponge_edge_centralities}
\alias{sponge_edge_centralities}
\title{Computes edge centralities}
\usage{
sponge_edge_centralities(sponge_result)
}
\arguments{
\item{sponge_result}{The output generated by the sponge method.}
}
\value{
data.table with gene, degree, eigenvector and betweenness
}
\description{
Computes edge betweenness
centrality for the ceRNA interaction network induced by the results of the
SPONGE method.
}
\examples{
sponge_edge_centralities(ceRNA_interactions)
}
\seealso{
sponge
}
