% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_network_centralities.R
\name{sponge_node_centralities}
\alias{sponge_node_centralities}
\title{Computes various node centralities}
\usage{
sponge_node_centralities(sponge_result, directed = FALSE)
}
\arguments{
\item{sponge_result}{output of the sponge method}
}
\value{
data.table with gene, degree, eigenvector and betweenness
}
\description{
Computes degree, eigenvector centrality and betweenness
centrality for the ceRNA interaction network induced by the results of the
SPONGE method
}
\seealso{
sponge
}

