% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_network_centralities.R
\name{sponge_node_centralities}
\alias{sponge_node_centralities}
\title{Computes various node centralities}
\usage{
sponge_node_centralities(sponge_result, directed = FALSE)
}
\arguments{
\item{sponge_result}{output of the sponge method}
}
\value{
data.table with gene, degree, eigenvector and betweenness
}
\description{
Computes degree, eigenvector centrality and betweenness
centrality for the ceRNA interaction network induced by the results of the
SPONGE method
}
\examples{
#First we extract miRNA candidates for each of the genes
genes_miRNA_candidates <- gene_miRNA_interaction_filter(
gene_expr = gene_expr,
mir_expr = mir_expr,
binding_db = NULL)

#Second we compute ceRNA interactions for all pairwise combinations of genes
#using all miRNAs remaining after filtering through elasticnet.
ceRNA_interactions <- sponge(
gene_expr = gene_expr,
mir_expr = mir_expr,
mir_interactions = genes_miRNA_candidates)

#Third, we compute the node centralities
sponge_node_centralities(ceRNA_interactions)
}
\seealso{
sponge
}

