% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{targetscan_ensg}
\alias{targetscan_ensg}
\title{targetscan predicted miRNA gene interactions}
\format{
A matrix gene ensembl ids vs miRNA family names. >=1 if interaction
is predicted, 0 otherwise
}
\source{
http://www.targetscan.org/vert_71/
}
\usage{
targetscan_ensg
}
\description{
targetscan predicted miRNA gene interactions
}
\keyword{datasets}
