% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_cancer_gene_expr}
\alias{test_cancer_gene_expr}
\title{example test expression data for spongEffects}
\format{
a matrix with gene expression data
}
\usage{
test_cancer_gene_expr
}
\description{
example test expression data for spongEffects
}
\keyword{datasets}
