% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_cancer_metadata}
\alias{test_cancer_metadata}
\title{example test sample meta data for spongEffects}
\format{
a data frame with sample meta data, SUBTYPE must be inside your dataframe
}
\usage{
test_cancer_metadata
}
\description{
example test sample meta data for spongEffects
}
\keyword{datasets}
