% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{train_cancer_metadata}
\alias{train_cancer_metadata}
\title{example training sample meta data for spongEffects}
\format{
a data frame with sample meta data, SUBTYPE must be inside your dataframe
}
\usage{
train_cancer_metadata
}
\description{
example training sample meta data for spongEffects
}
\keyword{datasets}
